add view claim for user

+++ src/main/java/com/entity/BankBill.java
@@ -66,7 +66,7 @@ public class BankBill {
         this.typeOfBankBill = typeOfBankBill;
     }
 
-    public int getNumberOfBill() {
+    public long getNumberOfBill() {
         return numberOfBill;
     }
